refactor: use uint16 for encoder config limits
Changes:
- Changed EncoderConfig fields to uint16: MaxObjectDepth, MaxStringSize, MaxContainerSize
- Updated bindings constants to uint16: MaxStringLength, MaxContainerSize, MaxMapDepth
- Removed unnecessary uint16 clamping in encoder.go since values are already uint16
- Added math import to waf_test.go
- Updated test files to cast uint16 to int where needed for loops and comparisons

This simplifies the code by using the correct types from the start, eliminating
redundant type conversions and clamping operations.

+++ encoder_decoder_test.go
@@ -660,9 +660,9 @@ func TestEncoderLimits(t *testing.T) {
 		encoder, err := newEncoder(EncoderConfig{
 			Pinner:           &pinner,
 			Timer:            encodeTimer,
-			MaxObjectDepth:   maxValueDepth,
-			MaxStringSize:    maxStringLength,
-			MaxContainerSize: maxContainerLength,
+			MaxObjectDepth:   uint16(maxValueDepth),
+			MaxStringSize:    uint16(maxStringLength),
+			MaxContainerSize: uint16(maxContainerLength),
 		})
 
 		require.NoError(t, err)
